+++ src/Handlers/TestHandler.php
@@ -24,29 +24,61 @@ class TestHandler extends ActionHandler
         return 'ok';
     }
 
+    /**
+     * Summary of meta
+     * @return null
+     */
     public function meta()
     {
+        // modify this to point to your book directory
         $dbPath = $this->dbConfig['db_path'];
         $epubPath = $this->dbConfig['epub_path'];
-        // modify this to point to your book directory
         $bookdir = $dbPath . DIRECTORY_SEPARATOR . $epubPath . DIRECTORY_SEPARATOR;
+
+        // allow ebooks in subfolders
+        $recursive = true;
         // baseurl for assets etc. (relative to this entrypoint)
         $endpoint = $this->request->getEndpoint();
         $basedir = dirname(str_replace('/index.php', '', $endpoint));
         $baseurl = $basedir . '/vendor/mikespub/php-epub-meta';
         // rename file as new "$author-$title.epub" after update
         $rename = false;
+        // template files directory
+        //$templatedir = dirname(__DIR__, 2) . '/vendor/mikespub/php-epub-meta/templates/';
+        $templatedir = dirname(__DIR__, 2) . '/templates/php-epub-meta/';
+        // cache directory for Google Books API calls (optional)
+        $cachedir = null;
+        if (!empty($this->cacheDir)) {
+            $cachedir = $this->cacheDir . '/google/titles/';
+        }
 
         // @todo let RequestHandler know we handled this
         putenv('PHPUNIT_TESTING=1');
 
-        $handler = new MetaAppHandler($bookdir, $baseurl, $rename);
+        // 1. create config array with the options above
+        $config = [
+            'bookdir' => $bookdir,
+            'recursive' => $recursive,
+            'baseurl' => $baseurl,
+            'rename' => $rename,
+            'templatedir' => $templatedir,
+            'cachedir' => $cachedir,
+        ];
+        // 2. instantiate the app handler with the config array
+        $handler = new MetaAppHandler($config);
         try {
-            $result = $handler->handle();
+            // 3. get request params from PHP globals or elsewhere
+            $params = $handler->getRequestFromGlobals();
+            // 4. handle the request and get back the result
+            $result = $handler->handle($params);
             if (!is_null($result)) {
+                // 5.a. output the result
                 echo $result;
+            } else {
+                // 5.b. some actions have no output, e.g. redirect after save
             }
         } catch (Throwable $e) {
+            // 6. catch any errors and handle them
             error_log($e);
             echo $e->getMessage();
         }